Extract the shared create-from-source logic
Closes #41.

"Create from Source" has two entry points - the button on Child items and the
"..." menu on the tree node - and each carried its own copy of the same 322
lines: scaffold from the blueprint, apply the workflow's mappings, create the
document, save it.

They had already drifted. The collection action never passed documentTypeName to
the modal, so the Destination tab showed an empty Document Type box when the
panel was opened from the button, and the correct name when opened from the
menu. Fixed here, since it is the same root cause.

createDocumentFromSource() in create-from-source.ts is now the single copy. It
takes fetch and the auth token as arguments rather than reaching for browser
globals, and touches no Umbraco contexts or repositories - so it runs equally in
the backoffice and in Node. That is what lets UpDoc's MCP server create
documents without driving a browser (#132), which is the reason this was worth
doing now rather than as tidying.

The UI stays with each caller: discovering allowed child types, opening the
pickers, notifications, navigating to the new document. Only the work is shared.

620 + 641 lines become 180 + 233 + 564.

Two things changed shape while moving, worth knowing when comparing:

- convertFieldTypes runs one loop with early continues rather than four loops
  over the same fields
- writeBlockProperty replaces two near-identical blocks in applyBlockGridValue
  and applyBlockValueByContentType

Behaviour is unchanged in both cases.

The e2e spec now points at the mirror's media layout - brochures sit flat in
Tailored Tours, not the old test site's PDF > {society} nesting. The mirror is
the model now.

Verified: TARGET=mirror create-from-source.spec.ts, 5/5 passing, including the
full create flow and the mapped preview against a real brochure.
